+++ src/cmd_runner/CmdGenInterface.ts
@@ -1,6 +1,0 @@
-'use strict';
-
-export interface CmdGenInterface {
-    list_file_cmd ();
-    build_database_cmd ();
-};